streamlit_app/requirements.txt
absl-py==1.4.0
altair==4.2.2
anyio==4.4.0
astor==0.8.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
easygui==0.98.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.9.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
img2table==1.2.11
imgaug==0.4.0
importlib-metadata==6.11.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.2.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==23.2
paddleocr==2.8.0
paddlepaddle==2.6.1
pandas==1.5.3
Pillow==9.5.0
polars==1.2.1
proto-plus==1.24.0
protobuf==3.19.0
pyarrow==16.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
python-bidi==0.5.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
rapidfuzz==3.9.4
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.0
semver==3.0.2
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.20.0
streamlit-aggrid==1.0.5
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.7.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
XlsxWriter==3.2.0
zipp==3.19.2
